Update unipn submodule and enhance deadlock analysis
- Updated the unipn submodule to commit 36e2bcb87e8061ce8ab0c9f8ba8fd33c485179a3.
- Added `find_deadlocks` function to improve deadlock detection clarity.
- Refactored `classify_counterexample` and related functions to utilize `CvnState` for better type safety and context in bug reports.
- Adjusted resource handling in the translator to use `usize` for token counts, enhancing memory efficiency.

+++ src/translator/goals.rs
@@ -58,7 +58,7 @@ pub fn translate_goals(program: &Program, net: &unipn::CvnNet) -> (Vec<GoalSpec>
         let mut predicates = Vec::new();
 
         for (key, count) in &goal.marking {
-            match marking_predicate(key, *count, &resource_by_name, &place_by_name) {
+            match marking_predicate(key, *count as usize, &resource_by_name, &place_by_name) {
                 Ok(pred) => predicates.push(pred),
                 Err(msg) => warnings.push(format!(
                     "goal '{}': marking key '{}' — {msg}",
